move pom file by new package
