+++ integration/src/test/java/com/tutrit/ApplicationRunTest.java
@@ -1,6 +1,7 @@
 package com.tutrit;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.tutrit.webClientRun.ApplicationRun;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
